+++ .gitignore
@@ -1,2 +1,4 @@
 Cargo.lock
 target
+*.bk
+*.orig
