+++ tests/consts/mod.rs
@@ -1,0 +1,65 @@
+#![allow(unused)]
+
+use num_bigint::BigDigit;
+
+pub const N1: BigDigit = -1i32 as BigDigit;
+pub const N2: BigDigit = -2i32 as BigDigit;
+
+pub const SUM_TRIPLES: &'static [(
+    &'static [BigDigit],
+    &'static [BigDigit],
+    &'static [BigDigit],
+)] = &[
+    (&[], &[], &[]),
+    (&[], &[1], &[1]),
+    (&[1], &[1], &[2]),
+    (&[1], &[1, 1], &[2, 1]),
+    (&[1], &[N1], &[0, 1]),
+    (&[1], &[N1, N1], &[0, 0, 1]),
+    (&[N1, N1], &[N1, N1], &[N2, N1, 1]),
+    (&[1, 1, 1], &[N1, N1], &[0, 1, 2]),
+    (&[2, 2, 1], &[N1, N2], &[1, 1, 2]),
+];
+
+pub const M: u32 = ::std::u32::MAX;
+pub const MUL_TRIPLES: &'static [(
+    &'static [BigDigit],
+    &'static [BigDigit],
+    &'static [BigDigit],
+)] = &[
+    (&[], &[], &[]),
+    (&[], &[1], &[]),
+    (&[2], &[], &[]),
+    (&[1], &[1], &[1]),
+    (&[2], &[3], &[6]),
+    (&[1], &[1, 1, 1], &[1, 1, 1]),
+    (&[1, 2, 3], &[3], &[3, 6, 9]),
+    (&[1, 1, 1], &[N1], &[N1, N1, N1]),
+    (&[1, 2, 3], &[N1], &[N1, N2, N2, 2]),
+    (&[1, 2, 3, 4], &[N1], &[N1, N2, N2, N2, 3]),
+    (&[N1], &[N1], &[1, N2]),
+    (&[N1, N1], &[N1], &[1, N1, N2]),
+    (&[N1, N1, N1], &[N1], &[1, N1, N1, N2]),
+    (&[N1, N1, N1, N1], &[N1], &[1, N1, N1, N1, N2]),
+    (&[M / 2 + 1], &[2], &[0, 1]),
+    (&[0, M / 2 + 1], &[2], &[0, 0, 1]),
+    (&[1, 2], &[1, 2, 3], &[1, 4, 7, 6]),
+    (&[N1, N1], &[N1, N1, N1], &[1, 0, N1, N2, N1]),
+    (&[N1, N1, N1], &[N1, N1, N1, N1], &[1, 0, 0, N1, N2, N1, N1]),
+    (&[0, 0, 1], &[1, 2, 3], &[0, 0, 1, 2, 3]),
+    (&[0, 0, 1], &[0, 0, 0, 1], &[0, 0, 0, 0, 0, 1]),
+];
+
+pub const DIV_REM_QUADRUPLES: &'static [(
+    &'static [BigDigit],
+    &'static [BigDigit],
+    &'static [BigDigit],
+    &'static [BigDigit],
+)] = &[
+    (&[1], &[2], &[], &[1]),
+    (&[3], &[2], &[1], &[1]),
+    (&[1, 1], &[2], &[M / 2 + 1], &[1]),
+    (&[1, 1, 1], &[2], &[M / 2 + 1, M / 2 + 1], &[1]),
+    (&[0, 1], &[N1], &[1], &[1]),
+    (&[N1, N1], &[N2], &[2, 1], &[3]),
+];
